wallet: fixed multiple issues with transaction inputs/outputs

+++ src/chain.h
@@ -62,7 +62,7 @@ int get_tx_key(uint8_t *buffer, uint8_t *tx_id);
 int get_unspent_tx_key(uint8_t *buffer, uint8_t *tx_id);
 int get_block_key(uint8_t *buffer, uint8_t *block_hash);
 
-uint32_t get_balance_for_address(uint8_t *address);
+uint64_t get_balance_for_address(uint8_t *address);
 
 #ifdef __cplusplus
 }
